add profile endpoint, rename spreadsheet to timesheet
